Handle vector variables
When taking the derivative wrt a vector variable, break the variable into components
and add those to the list of variables.
Adjust how the temporary names are created - if the variable is Indxed, dont' use the
full string - need to pull out the index base, create the temporary, and then take
the index.